chore: implement RouterProvider for routing management and hash fragment handling
