+++ apps/docs/src/components/router/index.ts
@@ -1,0 +1,2 @@
+export { RouterProvider } from "./router";
+export { default } from "./router";
